Re-arrange test cases
