+++ component/src/test/resources/testng.xml
@@ -24,7 +24,7 @@
     <test name="Siddhi-io-cdc-tests" enabled="true">
         <classes>
             <class name="org.wso2.extension.siddhi.io.cdc.source.TestCaseOfCDCSourceValidation"/>
-            <class name="org.wso2.extension.siddhi.io.cdc.source.TestCaseOfH2CDC"/>
+            <class name="org.wso2.extension.siddhi.io.cdc.source.TestCaseOfCDCPollingMode"/>
         </classes>
     </test>
 </suite>
